requirements.txt
flask-restx==0.5.1
Werkzeug==2.1.2
flask==2.1.2
gunicorn>=20.0,<21.0
pymongo[srv]==4.2.0
python-dotenv==0.21.0
autopep8==1.7.0